requirements.txt
biopython==1.74
Django==2.2.6
numpy==1.17.2
python-codon-tables==0.1.8
pytz==2019.3
sqlparse==0.3.0
